refactor: separate Presentation and Main layers folders to queries and commands
