+++ src/main/factories/controllers/category/commands/add-category-controller-factory.ts
@@ -1,6 +1,6 @@
 import { AddCategoryFactory } from '@/main/factories/usecases/category/commands'
 import { type Controller } from '@/presentation/protocols'
-import { AddCategoryController } from '@/presentation/controllers/category'
+import { AddCategoryController } from '@/presentation/controllers/category/commands'
 
 export class AddCategoryControllerFactory {
   public static makeAddCategoryController = (): Controller => {
