+++ src/main/factories/controllers/category/commands/index.ts
@@ -1,3 +1,2 @@
 export * from './add-category-controller-factory'
-export * from './get-all-categories-controller-factory'
 export * from './update-category-controller-factory'
