+++ src/main/factories/controllers/category/commands/update-category-controller-factory.ts
@@ -1,6 +1,6 @@
 import { UpdateCategoryFactory } from '@/main/factories/usecases/category/commands'
 import { type Controller } from '@/presentation/protocols'
-import { UpdateCategoryController } from '@/presentation/controllers/category/update-category-controller'
+import { UpdateCategoryController } from '@/presentation/controllers/category/commands/update-category-controller'
 
 export class UpdateCategoryControllerFactory {
   public static makeUpdateCategoryController = (): Controller => {
