+++ src/main/factories/controllers/category/queries/get-all-categories-controller-factory.ts
@@ -1,6 +1,6 @@
 import { GetAllCategoriesFactory } from '@/main/factories/usecases/category/queries'
 import { type Controller } from '@/presentation/protocols'
-import { GetAllCategoriesController } from '@/presentation/controllers/category'
+import { GetAllCategoriesController } from '@/presentation/controllers/category/queries'
 
 export class GetAllCategoriesControllerFactory {
   public static makeGetAllCategoriesController = (): Controller => {
