+++ src/main/factories/controllers/category/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './get-all-categories-controller-factory'
