+++ src/main/factories/controllers/product/commands/add-product-controller-factory.ts
@@ -1,6 +1,6 @@
 import { AddProductFactory } from '@/main/factories/usecases/product/commands'
 import { type Controller } from '@/presentation/protocols'
-import { AddProductController } from '@/presentation/controllers/product'
+import { AddProductController } from '@/presentation/controllers/product/commands'
 
 export class AddProductControllerFactory {
   public static makeAddProductController = (): Controller => {
