+++ src/presentation/controllers/category/commands/index.ts
@@ -1,0 +1,2 @@
+export * from './add-category-controller'
+export * from './update-category-controller'
