+++ src/presentation/controllers/category/queries/get-all-categories-controller.ts
