+++ src/presentation/controllers/category/queries/index.ts
@@ -1,2 +1,1 @@
-export * from './add-category-controller'
 export * from './get-all-categories-controller'
