+++ tests/presentation/controllers/category/commands/add-category-controller.spec.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 import { throwError } from '@/tests/test-helper'
 import { AddCategorySpy } from '@/tests/presentation/mocks/category'
 import { mockAddCategoryInput } from '@/tests/domain/mocks/category'
-import { AddCategoryController } from '@/presentation/controllers/category'
+import { AddCategoryController } from '@/presentation/controllers/category/commands'
 import { HttpHelper } from '@/presentation/helpers'
 import { EntityValidationError } from '@/domain/errors'
 import { CategoryAlreadyExistsError } from '@/domain/errors/category'
