+++ tests/presentation/controllers/product/commands/add-product-controller.spec.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 import { throwError } from '@/tests/test-helper'
 import { AddProductSpy } from '@/tests/presentation/mocks/product'
 import { mockAddProductInput } from '@/tests/domain/mocks/product'
-import { AddProductController } from '@/presentation/controllers/product'
+import { AddProductController } from '@/presentation/controllers/product/commands'
 import { HttpHelper } from '@/presentation/helpers'
 import { EntityValidationError } from '@/domain/errors'
 import { ProductAlreadyExistsError } from '@/domain/errors/product'
