+++ tests/presentation/controllers/product/commands/update-product-controller.spec.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 import { throwError } from '@/tests/test-helper'
 import { UpdateProductSpy } from '@/tests/presentation/mocks/product'
 import { mockUpdateProductInput } from '@/tests/domain/mocks/product'
-import { UpdateProductController } from '@/presentation/controllers/product'
+import { UpdateProductController } from '@/presentation/controllers/product/commands'
 import { HttpHelper } from '@/presentation/helpers'
 import { EntityValidationError } from '@/domain/errors'
 import { ProductNotFoundError } from '@/domain/errors/product'
